Communication interface for ROS edge broadcaster
