+++ supplementary/alica_msgs/msg/AgentAnnouncementList.msg
@@ -1,0 +1,1 @@
+AgentAnnouncement[] agent_announcements
