+++ supplementary/alica_msgs/msg/AgentQueryList.msg
@@ -1,0 +1,1 @@
+AgentQuery[] agent_queries
