+++ supplementary/alica_msgs/msg/PlanTreeInfoList.msg
@@ -1,0 +1,1 @@
+PlanTreeInfo[] plan_tree_infos
