+++ supplementary/alica_msgs/msg/SolverResultList.msg
@@ -1,0 +1,1 @@
+SolverResult[] solver_results
