+++ supplementary/alica_msgs/msg/SyncReadyList.msg
@@ -1,0 +1,1 @@
+SyncReady[] sync_readies
