+++ supplementary/alica_msgs/msg/SyncTalkList.msg
@@ -1,0 +1,1 @@
+SyncTalk[] sync_talks
