+++ supplementary/alica_ros1/alica_ros_proxy/include/communication/AlicaRosCommunication.h
@@ -1,84 +1,31 @@
 #pragma once
-#include "alica_msgs/AgentAnnouncement.h"
-#include "alica_msgs/AgentQuery.h"
-#include "alica_msgs/AlicaEngineInfo.h"
-#include "alica_msgs/AllocationAuthorityInfo.h"
-#include "alica_msgs/PlanTreeInfo.h"
-#include "alica_msgs/RoleSwitch.h"
-#include "alica_msgs/SolverResult.h"
-#include "alica_msgs/SyncReady.h"
-#include "alica_msgs/SyncTalk.h"
 
-#include <engine/IAlicaCommunication.h>
+#include "communication/AlicaRosCommunicationCommon.h"
 
 #include <ros/callback_queue.h>
 #include <ros/ros.h>
-#include <string>
 
 using namespace alica;
 
 namespace alicaRosProxy
 {
 
-class AlicaRosCommunication : public alica::IAlicaCommunication
+class AlicaRosCommunication : public AlicaRosCommunicationCommon
 {
 public:
     AlicaRosCommunication(const AlicaCommunicationHandlers& callbacks, ros::CallbackQueue& cb_queue = *ros::getGlobalCallbackQueue());
     virtual ~AlicaRosCommunication();
 
-    void tick() override;
-
-    void sendAllocationAuthority(const AllocationAuthorityInfo& aai) const override;
-    void sendAlicaEngineInfo(const AlicaEngineInfo& bi) const override;
-    void sendPlanTreeInfo(const PlanTreeInfo& pti) const override;
-    void sendRoleSwitch(const RoleSwitch& rs, AgentId agentID) const override;
-    void sendSyncReady(const SyncReady& sr) const override;
-    void sendSyncTalk(const SyncTalk& st) const override;
-    void sendSolverResult(const SolverResult& sr) const override;
-    void sendAgentQuery(const AgentQuery& pq) const override;
-    void sendAgentAnnouncement(const AgentAnnouncement& pa) const override;
-    void sendLogMessage(int level, const std::string& message) const override;
-
-    void handleAllocationAuthorityRos(const alica_msgs::AllocationAuthorityInfo& aai);
-    void handlePlanTreeInfoRos(alica_msgs::PlanTreeInfoPtr pti);
-    void handleSyncReadyRos(alica_msgs::SyncReadyPtr sr);
-    void handleSyncTalkRos(alica_msgs::SyncTalkPtr st);
-    void handleSolverResult(const alica_msgs::SolverResult& sr);
-    void handleAgentQuery(const alica_msgs::AgentQuery& pq);
-    void handleAgentAnnouncement(const alica_msgs::AgentAnnouncement& pa);
-
-    void startCommunication() override;
-    void stopCommunication() override;
-
 private:
-    ros::CallbackQueue& _callbackQueue;
-
-    ros::Publisher _alicaEngineInfoPublisher;
-    ros::Publisher _roleSwitchPublisher;
-
-    ros::Publisher _allocationAuthorityInfoPublisher;
-    ros::Subscriber _allocationAuthorityInfoSubscriber;
-
-    ros::Publisher _planTreeInfoPublisher;
-    ros::Subscriber _planTreeInfoSubscriber;
-
-    ros::Publisher _syncReadyPublisher;
-    ros::Subscriber _syncReadySubscriber;
-
-    ros::Publisher _syncTalkPublisher;
-    ros::Subscriber _syncTalkSubscriber;
-
-    ros::Publisher _solverResultPublisher;
-    ros::Subscriber _solverResultSubscriber;
-
-    ros::Publisher _presenceQueryPublisher;
-    ros::Subscriber _presenceQuerySubscriber;
-
-    ros::Publisher _presenceAnnouncementPublisher;
-    ros::Subscriber _presenceAnnouncementSubscriber;
-
     ros::NodeHandle _nh;
-    bool _isRunning;
+
+    ros::Subscriber _syncReadySubscriber;
+    ros::Subscriber _planTreeInfoSubscriber;
+    ros::Subscriber _solverResultSubscriber;
+    ros::Subscriber _presenceAnnouncementSubscriber;
+    ros::Subscriber _allocationAuthorityInfoSubscriber;
+    ros::Subscriber _syncTalkSubscriber;
+    ros::Subscriber _presenceQuerySubscriber;
 };
 
 } /* namespace alicaRosProxy */
